# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hub William API",
    "description": "Control-plane and gateway API for Hub William agent pools",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/agent-connections": {
      "get": {
        "tags": ["agent connections"],
        "operationId": "list_connections",
        "responses": {
          "200": {
            "description": "Current user's provider connections",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AgentConnection"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Hub login required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/agent-connections/deepseek": {
      "post": {
        "tags": ["agent connections"],
        "operationId": "connect_deepseek",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectDeepseekRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "DeepSeek API key validated and encrypted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentConnection"
                }
              }
            }
          },
          "401": {
            "description": "Hub login required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid DeepSeek API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "DeepSeek validation unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/agent-connections/groq": {
      "post": {
        "tags": ["agent connections"],
        "operationId": "connect_groq",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectGroqRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentConnection"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/agent-connections/start": {
      "post": {
        "tags": ["agent connections"],
        "operationId": "start",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartAgentConnectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Provider authorization started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentConnection"
                }
              }
            }
          },
          "401": {
            "description": "Hub login required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Provider authorization unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/agent-connections/{connection_id}": {
      "get": {
        "tags": ["agent connections"],
        "operationId": "get_connection",
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "description": "Agent connection ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Latest provider connection status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentConnection"
                }
              }
            }
          },
          "404": {
            "description": "Connection not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["agent connections"],
        "operationId": "disconnect",
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "description": "Agent connection ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Connected account pool deleted"
          },
          "404": {
            "description": "Connection not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/agent-connections/{connection_id}/complete": {
      "post": {
        "tags": ["agent connections"],
        "operationId": "complete_authorization",
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "description": "Agent connection ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteAuthorizationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Authorization completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentConnection"
                }
              }
            }
          },
          "422": {
            "description": "Invalid callback URL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Provider rejected the exchange",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/agent-connections/{connection_id}/refresh": {
      "post": {
        "tags": ["agent connections"],
        "operationId": "refresh_connection",
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "description": "Agent connection ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Credential refreshed or provider reauthorization started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentConnection"
                }
              }
            }
          },
          "401": {
            "description": "Hub login required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Connection not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Provider refresh unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/agent-pool-requests/{request_id}/decision": {
      "post": {
        "tags": ["agent pools"],
        "operationId": "decide_request",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "description": "Join request identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecideAgentPoolJoinRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Join request decided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentPoolJoinRequest"
                }
              }
            }
          },
          "403": {
            "description": "Pool ownership required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/agent-pools/{connection_id}/members": {
      "post": {
        "tags": ["agent pools"],
        "operationId": "invite_member",
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "description": "Connected account identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteAgentPoolMember"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Member invited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentPoolPerson"
                }
              }
            }
          },
          "403": {
            "description": "Pool ownership required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invite rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/agent-pools/{connection_id}/members/{username}": {
      "delete": {
        "tags": ["agent pools"],
        "operationId": "remove_member",
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "description": "Connected account identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "username",
            "in": "path",
            "description": "Hub William username",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Member removed"
          },
          "403": {
            "description": "Pool ownership required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Membership not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/refresh": {
      "post": {
        "tags": ["auth"],
        "operationId": "refresh",
        "responses": {
          "200": {
            "description": "Access and refresh sessions rotated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Refresh session missing or expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/organization-invitations/{invitation_id}": {
      "delete": {
        "tags": ["organizations"],
        "operationId": "decline_invitation",
        "parameters": [
          {
            "name": "invitation_id",
            "in": "path",
            "description": "Pending membership ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Invitation declined"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentConnection": {
        "type": "object",
        "required": [
          "availability_status",
          "created_at",
          "id",
          "provider",
          "status",
          "updated_at"
        ],
        "properties": {
          "account_label": {
            "type": ["string", "null"]
          },
          "authorization": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AgentAuthorizationPrompt"
              }
            ]
          },
          "availability_status": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "failure_message": {
            "type": ["string", "null"]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "plan": {
            "type": ["string", "null"]
          },
          "provider": {
            "$ref": "#/components/schemas/AgentProvider"
          },
          "status": {
            "$ref": "#/components/schemas/AgentConnectionStatus"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": ["code", "message"],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}