# apps/api/src/lib.rs
mod agent_pools;
mod auth;
mod config;
mod connections;
mod error;
mod gateway;
mod health;
mod openapi;
mod organizations;
mod playground;
mod playground_groq_voice;
mod playground_voice;
mod provider_catalogue;
mod provider_catalogue_generated;
mod telegram;
mod telegram_catalogue;
mod usage;

use std::time::Duration;

use axum::{
    Router,
    extract::DefaultBodyLimit,
    http::{
        Method,
        header::{AUTHORIZATION, CONTENT_TYPE},
    },
    routing::{get, post, put},
};
use reqwest::Client;
use sqlx::PgPool;
use tower_http::cors::AllowOrigin;
use tower_http::cors::CorsLayer;
use utoipa::OpenApi;
use utoipa_swagger_ui::SwaggerUi;

pub use agent_pools::{
    AgentPool, AgentPoolAvailability, AgentPoolAvailabilityStatus, AgentPoolJoinRequest,
    AgentPoolPerson, AgentPoolRequestStatus, AgentPoolShareEvidence, AgentPoolUsageMetric,
    CreateAgentPoolJoinRequest, DecideAgentPoolJoinRequest, InviteAgentPoolMember,
};
use agent_pools::{
    create_request, decide_request, invite_member, list as list_agent_pools, remove_member,
};
pub use auth::{AuthenticatedUser, LoginRequest, RegisterRequest, SessionResponse};
use auth::{login, logout, refresh, register, session};
pub use config::AppConfig;
pub use connections::{
    AgentConnection, AgentConnectionStatus, AgentProvider, CompleteAuthorizationRequest,
    ConnectDeepseekRequest, ConnectGroqRequest, ProviderCredentialRefreshResult,
    ProviderCredentialRefreshStatus, ProviderCredentialRefreshSummary, StartAgentConnectionRequest,
    refresh_all_provider_credentials, refresh_due_provider_credentials,
    refresh_nightly_provider_credentials, refresh_stored_connection_metadata,
};
use connections::{
    complete_authorization, connect_deepseek, connect_groq, disconnect, get_connection,
    list_connections, refresh_connection, start,
};
pub use error::ErrorResponse;
pub use gateway::{CreatedGatewayKey, GatewayKey};
use gateway::{
    claude_count_tokens, claude_messages, claude_models, create_key, deepseek_chat,
    deepseek_models, deepseek_responses, gemini_models, gemini_request, grok_chat, grok_models,
    grok_responses, list_keys, openai_models, openai_responses, revoke_key,
};
pub use health::HealthResponse;
use health::health;
pub use openapi::ApiDoc;
pub use organizations::{
    CreateOrganization, InviteOrganizationMember, Organization, OrganizationAgent,
    OrganizationAgentDetails, OrganizationAgentListQuery, OrganizationInvitation,
    OrganizationMember, OrganizationOverview, OrganizationPeriodQuery, OrganizationUsage,
    OrganizationUsageBreakdown, OrganizationUsageDay, OrganizationUsageQuery,
    ShareOrganizationAgent,
};
pub use telegram::{
    CreateTelegramOrder, SepayResult, SepayTransaction, TelegramAudience, TelegramOrder,
};
use telegram::{
    block_contact, cancel_order, create_order, get_contact, get_order, list_audience, list_orders,
    observe_contact, record_sepay_payment, set_contact_language,
};
pub use telegram_catalogue::{
    AdjustTelegramProduct, CreateTelegramProduct, RestockTelegramProduct, TelegramCatalogue,
    TelegramCatalogueProvider, TelegramProduct, TelegramRestock,
};
use telegram_catalogue::{
    adjust_product, catalogue, create_product, full_catalogue, get_product, restock_product,
};

#[derive(Clone)]
pub struct AppState {
    pub config: AppConfig,
    pub http: Client,
    pub gateway_http: Client,
    pub pool: PgPool,
}

pub fn gateway_http_client() -> Result<Client, reqwest::Error> {
    Client::builder()
        .connect_timeout(Duration::from_secs(10))
        .user_agent(concat!("hub-william/", env!("CARGO_PKG_VERSION")))
        .build()
}

fn browser_origins(config: &AppConfig) -> Vec<axum::http::HeaderValue> {
    let mut browser_origins = vec![config.frontend_origin.clone()];
    if !config.cookie_secure {
        for origin in [
            "http://localhost:5173".parse().expect("valid local origin"),
            "http://127.0.0.1:5173".parse().expect("valid local origin"),
            "http://localhost:5174".parse().expect("valid local origin"),
            "http://127.0.0.1:5174".parse().expect("valid local origin"),
            "http://localhost:3000".parse().expect("valid local origin"),
            "http://127.0.0.1:3000".parse().expect("valid local origin"),
        ] {
            if !browser_origins.contains(&origin) {
                browser_origins.push(origin);
            }
        }
    }
    browser_origins
}

pub fn app(state: AppState) -> Router {
    let cors = CorsLayer::new()
        .allow_origin(AllowOrigin::list(browser_origins(&state.config)))
        .allow_methods([Method::GET, Method::POST, Method::DELETE])
        .allow_headers([AUTHORIZATION, CONTENT_TYPE])
        .expose_headers([axum::http::HeaderName::from_static(
            gateway::SERVED_CONNECTION_HEADER,
        )])
        .allow_credentials(true);

    let gateway_routes = Router::new()
        .route("/gateway/openai/v1/responses", post(openai_responses))
        .route("/gateway/openai/v1/models", get(openai_models))
        .route("/gateway/claude/v1/messages", post(claude_messages))
        .route("/gateway/claude/v1/models", get(claude_models))
        .route(
            "/gateway/claude/v1/messages/count_tokens",
            post(claude_count_tokens),
        )
        .route("/gateway/grok/v1/chat/completions", post(grok_chat))
        .route("/gateway/grok/v1/responses", post(grok_responses))
        .route("/gateway/grok/v1/models", get(grok_models))
        .route("/gateway/gemini/v1beta/models", get(gemini_models))
        .route(
            "/gateway/gemini/v1beta/models/{*operation}",
            post(gemini_request),
        )
        .route("/gateway/deepseek/chat/completions", post(deepseek_chat))
        .route("/gateway/deepseek/responses", post(deepseek_responses))
        .route("/gateway/deepseek/models", get(deepseek_models))
        .route("/gateway/groq/v1/models", get(gateway::groq_models))
        .route("/gateway/groq/v1/responses", post(gateway::groq_responses))
        .route(
            "/gateway/groq/v1/chat/completions",
            post(gateway::groq_chat),
        )
        .layer(DefaultBodyLimit::disable());

    Router::new()
        .route("/health", get(health))
        .route("/auth/register", post(register))
        .route("/auth/login", post(login))
        .route("/auth/session", get(session))
        .route("/auth/logout", post(logout))
        .route("/auth/refresh", post(refresh))
        .route("/internal/telegram/contacts", post(observe_contact))
        .route(
            "/internal/telegram/contacts/{telegram_user_id}",
            get(get_contact),
        )
        .route(
            "/internal/telegram/contacts/{telegram_user_id}/language",
            put(set_contact_language),
        )
        .route(
            "/internal/telegram/contacts/{telegram_user_id}/orders",
            get(list_orders).post(create_order),
        )
        .route(
            "/internal/telegram/contacts/{telegram_user_id}/orders/{reference}",
            get(get_order),
        )
        .route(
            "/internal/telegram/contacts/{telegram_user_id}/orders/{reference}/cancel",
            post(cancel_order),
        )
        .route(
            "/internal/telegram/payments/sepay",
            post(record_sepay_payment),
        )
        .route("/internal/telegram/audience", get(list_audience))
        .route(
            "/internal/telegram/audience/{chat_id}/block",
            post(block_contact),
        )
        .route("/internal/telegram/catalogue", get(catalogue))
        .route("/internal/telegram/catalogue/full", get(full_catalogue))
        .route("/internal/telegram/products", post(create_product))
        .route(
            "/internal/telegram/products/{slug}",
            get(get_product).patch(adjust_product),
        )
        .route(
            "/internal/telegram/products/{slug}/restock",
            post(restock_product),
        )
        .route("/agent-pools", get(list_agent_pools))
        .route(
            "/organizations",
            get(organizations::list_organizations).post(organizations::create_organization),
        )
        .route(
            "/organization-invitations",
            get(organizations::list_invitations),
        )
        .route(
            "/organization-invitations/{invitation_id}/accept",
            post(organizations::accept_invitation),
        )
        .route(
            "/organization-invitations/{invitation_id}",
            axum::routing::delete(organizations::decline_invitation),
        )
        .route("/organizations/{id}/overview", get(organizations::overview))
        .route(
            "/organizations/{id}/agents",
            get(organizations::list_agents).post(organizations::share_agent),
        )
        .route(
            "/organizations/{id}/agents/{connection_id}",
            axum::routing::delete(organizations::unshare_agent),
        )
        .route(
            "/organizations/{id}/members",
            get(organizations::list_members).post(organizations::invite_member),
        )
        .route(
            "/organizations/{id}/members/{username}",
            axum::routing::delete(organizations::remove_member),
        )
        .route("/organizations/{id}/usage", get(organizations::usage))
        .route(
            "/agent-pools/{connection_id}/requests",
            post(create_request),
        )
        .route("/agent-pools/{connection_id}/members", post(invite_member))
        .route(
            "/agent-pools/{connection_id}/members/{username}",
            axum::routing::delete(remove_member),
        )
        .route(
            "/agent-pool-requests/{request_id}/decision",
            post(decide_request),
        )
        .route("/agent-connections", get(list_connections))
        .route("/agent-connections/deepseek", post(connect_deepseek))
        .route("/agent-connections/groq", post(connect_groq))
        .route("/agent-connections/start", post(start))
        .route(
            "/agent-connections/{connection_id}",
            get(get_connection).delete(disconnect),
        )
        .route(
            "/agent-connections/{connection_id}/complete",
            post(complete_authorization),
        )
        .route(
            "/agent-connections/{connection_id}/refresh",
            post(refresh_connection),
        )
        .route("/gateway-keys", get(list_keys).post(create_key))
        .route("/gateway-keys/{key_id}", axum::routing::delete(revoke_key))
        .route(
            "/playground/{provider}/accounts/{connection_id}/models",
            get(playground::models),
        )
        .merge(
            Router::new()
                .route("/playground/chat", post(playground::chat))
                .route(
                    "/playground/groq/voice/turn",
                    post(playground_groq_voice::create).layer(DefaultBodyLimit::max(64 * 1024)),
                )
                .route(
                    "/playground/voice",
                    post(playground_voice::create).layer(DefaultBodyLimit::max(96 * 1024)),
                )
                .layer(DefaultBodyLimit::max(32 * 1024 * 1024))
                .layer(axum::middleware::from_fn_with_state(
                    state.clone(),
                    playground::authorize_chat,
                )),
        )
        .merge(gateway_routes)
        .merge(SwaggerUi::new("/docs").url("/api-docs/openapi.json", ApiDoc::openapi()))
        .layer(cors)
        .with_state(state)
}

#[cfg(test)]
mod tests {
    use axum::{
        body::{Body, to_bytes},
        http::{Request, StatusCode},
    };
    use reqwest::Client;
    use tower::ServiceExt;

    use sqlx::postgres::PgPoolOptions;

    use super::{AppConfig, AppState, app, gateway_http_client};

    #[tokio::test]
    async fn health_endpoint_identifies_the_service() {
        let state = AppState {
            config: AppConfig::default(),
            http: Client::new(),
            gateway_http: Client::new(),
            pool: PgPoolOptions::new()
                .connect_lazy("postgres://localhost/hub_william_test")
                .expect("test database URL should parse"),
        };
        let response = app(state)
            .oneshot(
                Request::builder()
                    .uri("/health")
                    .body(Body::empty())
                    .expect("health request should be valid"),
            )
            .await
            .expect("health route should respond");

        assert_eq!(response.status(), StatusCode::OK);

        let body = to_bytes(response.into_body(), usize::MAX)
            .await
            .expect("health response should have a body");
        let payload: serde_json::Value =
            serde_json::from_slice(&body).expect("health response should be valid JSON");

        assert_eq!(payload["status"], "ok");
        assert_eq!(payload["service"], "hub-william-backend");
    }

    #[tokio::test]
    async fn only_gateway_routes_accept_large_request_bodies() {
        let state = AppState {
            config: AppConfig::default(),
            http: Client::new(),
            gateway_http: Client::new(),
            pool: PgPoolOptions::new()
                .connect_lazy("postgres://localhost/hub_william_test")
                .expect("test database URL should parse"),
        };
        let gateway_body = vec![b' '; 17 * 1024 * 1024];

        let gateway_response = app(state.clone())
            .oneshot(
                Request::builder()
                    .method("POST")
                    .uri("/gateway/deepseek/responses")
                    .header("content-type", "application/json")
                    .body(Body::from(gateway_body))
                    .expect("large gateway request should build"),
            )
            .await
            .expect("gateway should respond");
        assert_eq!(gateway_response.status(), StatusCode::UNAUTHORIZED);

        let browser_response = app(state)
            .oneshot(
                Request::builder()
                    .method("POST")
                    .uri("/auth/login")
                    .header("content-type", "application/json")
                    .body(Body::from(vec![b' '; 3 * 1024 * 1024]))
                    .expect("large browser request should build"),
            )
            .await
            .expect("browser API should respond");
        assert_eq!(browser_response.status(), StatusCode::PAYLOAD_TOO_LARGE);
    }

    #[test]
    fn gateway_http_client_has_no_total_request_timeout() {
        let client = gateway_http_client().expect("gateway client should build");
        let debug = format!("{client:?}");

        assert!(
            !debug.contains("TotalTimeout"),
            "gateway client unexpectedly has a total request timeout: {debug}"
        );
    }

    #[tokio::test]
    async fn gemini_native_model_routes_reach_gateway_authentication() {
        let state = AppState {
            config: AppConfig::default(),
            http: Client::new(),
            gateway_http: Client::new(),
            pool: PgPoolOptions::new()
                .connect_lazy("postgres://localhost/hub_william_test")
                .expect("test database URL should parse"),
        };
        let response = app(state)
            .oneshot(
                Request::builder()
                    .method("POST")
                    .uri("/gateway/gemini/v1beta/models/gemini-test:generateContent")
                    .header("content-type", "application/json")
                    .header("x-goog-api-key", "invalid")
                    .body(Body::from(r#"{"contents": []}"#))
                    .expect("Gemini request should be valid"),
            )
            .await
            .expect("Gemini route should respond");

        assert_eq!(response.status(), StatusCode::UNAUTHORIZED);
    }

    #[tokio::test]
    async fn local_cors_allows_only_the_supported_development_origins() {
        let config = AppConfig {
            cookie_secure: false,
            ..AppConfig::default()
        };
        let state = AppState {
            config,
            http: Client::new(),
            gateway_http: Client::new(),
            pool: PgPoolOptions::new()
                .connect_lazy("postgres://localhost/hub_william_test")
                .expect("test database URL should parse"),
        };
        let service = app(state);

        for origin in [
            "http://localhost:5173",
            "http://127.0.0.1:5173",
            "http://localhost:5174",
            "http://127.0.0.1:5174",
            "http://localhost:3000",
            "http://127.0.0.1:3000",
        ] {
            let response = service
                .clone()
                .oneshot(
                    Request::builder()
                        .method("OPTIONS")
                        .uri("/agent-pools")
                        .header("origin", origin)
                        .header("access-control-request-method", "GET")
                        .body(Body::empty())
                        .expect("CORS preflight request should be valid"),
                )
                .await
                .expect("CORS preflight should respond");

            assert_eq!(response.status(), StatusCode::OK);
            assert_eq!(response.headers()["access-control-allow-origin"], origin);
        }

        let response = service
            .oneshot(
                Request::builder()
                    .method("OPTIONS")
                    .uri("/agent-pools")
                    .header("origin", "http://localhost:9999")
                    .header("access-control-request-method", "GET")
                    .body(Body::empty())
                    .expect("unsupported-origin request should be valid"),
            )
            .await
            .expect("unsupported-origin preflight should respond");

        assert!(
            response
                .headers()
                .get("access-control-allow-origin")
                .is_none()
        );
    }
}